+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `KPInputOTPField` helper and error text's maxLines changed from `Int.MAX_VALUE` to `2`.
+- `KPInputPhoneNumberField` error text's maxLines changed from `Int.MAX_VALUE` to `2`.
 
 ## [0.16.0] - 2025-01-13